feat(folder): Added grid parsing functionality to Folder class
Right now, the information about the grids has to be parsed at a recipe level each time. I have added a couple of harmless parsing functions to that will help in consolidating the grid data for a project and reduce code duplication. The parsing functions themselves were added in the gridutil module. Also corrected a minor docstring error.