fix(gridutil): Allow .ill files with headers to be restored
The fact that I could not pass .ill files with headers to this function meant that I couldn't use it correctly in the comfort maps and annual-irradaince recipes, where we process the direct results separately from the total. This allows the i'll files with headers to be restructured without changing the runtime significantly.