perf(combined-receiver-output): Allow the auto vmtx path to be optional
The combined receiver functionality that exists is pretty handy. However, the redirecting of view matrix paths automatically is a problem. The issues are several: firstly, the "o=%s..%s.vmx" does not allow any flexibility in where the view matrix would written to and writing to the root directory itself will lead to confusion for large projects with several view matrices littering the root folder. Secondly, this way of generating view matrices is somewhat of a black box as one as to write additional logic to locate the view matrices that were generated automatically. So, I am allowing an opt-out feature such that the view matrices are not automatically redirected (if one chooses to do so).